+++ src/gr.c
@@ -2205,7 +2205,7 @@ void DrawSolidColourSpriteScaled(Viewport *viewport, short x, short y,
 }
 
 /* Function start: 0x4283A4 */
-int GetTransformedShapeBounds(Viewport *viewport, short x, short y,
+short GetTransformedShapeBounds(Viewport *viewport, short x, short y,
                               unsigned char *shape, short frame,
                               short angle, short scale, short flip,
                               short *bounds)
